refactor: split turn resolver test harness into domain peers (#4252)
Slice A: replace the 632-line god harness with combat, diplomacy, economy,
naval, and common peer modules; keep turn_resolver_test_harness.dart as a
thin re-export barrel with resolveTurnComplete.

Refs #4252
